requirement.txt.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow


//admin
//login:Nimap
//password:123